Rent-management-System/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
certifi==2023.5.7
cloudinary==1.33.0
confusable-homoglyphs==3.2.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.2
django-bootstrap4==23.1
django-cors-headers==4.3.1
django-registration==3.3
djangorestframework==3.14.0
filelock==3.12.0
gunicorn==20.1.0
Markdown==3.5.1
mysqlclient==2.1.1
platformdirs==3.5.1
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.0
